pyproject.toml
[tool.poetry]
name = "jarvis"
version = "0.1.0"
description = ""
authors = ['Teng "Niel" Hu <hu.niel92@gmail.com>']
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
openai = "1.2.0"
actionweaver = "^0.0.14"
langchain = "^0.0.285"
google-api-python-client = "^2.98.0"
html2text = "^2020.1.16"
streamlit = "^1.26.0"
# streamlit-pills = "^0.3.0"
# streamlit-webrtc = "0.47.1"
streamlit-audiorecorder= "0.0.4"
chromadb = "^0.4.17"

[tool.poetry.group.dev.dependencies]
jupyterlab = "^4.0.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
